Moved out certain API calls to debug.py & public.py
